CTCP-2358: Add authentication and message unwrapping
Note that I have removed the message size action. This is intentional, while we only expect 5MB to come back up through the small message route, we shouldn't reject anything that comes through from EIS.

I have also dropped the hybrid memory/file streaming to go for a pure file based streaming, due to the problem that Akka has some weird race conditions that they use Thread.sleep to work around in tests, so our code is actually too fast for Akka. We value reliability over speed, so we have done this.

We also have to consider the wrapped XML that EIS send. While I have written to our spec, I'm concerned that based on the us to EIS spec that we have to wrap the XML differently and that EIS will use the same wrapping. If so, the wrapping logic is now abstracted out so we just need to update the transformer -- the rest of our code works on the unwrapped XML.